pyproject.toml
[project]
name = "gmail-automation"
version = "0.1.0"
description = "Gmail メール自動PDF化システム"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-cloud-pubsub>=2.18.0",
    "weasyprint>=60.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gmail_automation"]
